Split up linear into pieces
