+++ examples/auto_table/auto_table.cc
@@ -1,5 +1,5 @@
 #include <gcs/constraints/all_different.hh>
-#include <gcs/constraints/linear_equality.hh>
+#include <gcs/constraints/linear.hh>
 #include <gcs/presolvers/auto_table.hh>
 #include <gcs/problem.hh>
 #include <gcs/search_heuristics.hh>
