+++ examples/cake/cake.cc
@@ -1,4 +1,3 @@
-#include <gcs/constraints/linear_equality.hh>
 #include <gcs/problem.hh>
 #include <gcs/search_heuristics.hh>
 #include <gcs/solve.hh>
