+++ examples/knapsack/knapsack.cc
@@ -1,5 +1,5 @@
 #include <gcs/constraints/knapsack.hh>
-#include <gcs/constraints/linear_equality.hh>
+#include <gcs/constraints/linear.hh>
 #include <gcs/problem.hh>
 #include <gcs/search_heuristics.hh>
 #include <gcs/solve.hh>
