+++ examples/money/money.cc
@@ -1,5 +1,5 @@
 #include <gcs/constraints/all_different.hh>
-#include <gcs/constraints/linear_equality.hh>
+#include <gcs/constraints/linear.hh>
 #include <gcs/problem.hh>
 #include <gcs/solve.hh>
