+++ gcs/constraints/linear/linear_test.cc
@@ -1,5 +1,5 @@
 #include <gcs/constraints/constraints_test_utils.hh>
-#include <gcs/constraints/linear_equality.hh>
+#include <gcs/constraints/linear.hh>
 #include <gcs/problem.hh>
 #include <gcs/solve.hh>
 
@@ -70,7 +70,7 @@ auto run_linear_test(bool proofs, const string & mode, pair<int, int> v1_range, 
 
     auto proof_name = proofs ? make_optional("linear_equality_test") : nullopt;
 
-    if ((! is_same_v<Constraint_, LinearEquality>) && 1 == ineqs.size())
+    if ((! is_same_v<Constraint_, LinearEquality>)&&1 == ineqs.size())
         solve_for_tests_checking_consistency(p, proof_name, expected, actual, tuple{pair{v1, CheckConsistency::BC}, pair{v2, CheckConsistency::BC}, pair{v3, CheckConsistency::BC}});
     else
         solve_for_tests(p, proof_name, actual, tuple{v1, v2, v3});
