+++ minicp_benchmarks/magic_series/magic_series.cc
@@ -1,5 +1,5 @@
 #include <gcs/constraints/equals.hh>
-#include <gcs/constraints/linear_equality.hh>
+#include <gcs/constraints/linear.hh>
 #include <gcs/problem.hh>
 #include <gcs/search_heuristics.hh>
 #include <gcs/solve.hh>
