+++ minicp_benchmarks/magic_square/magic_square.cc
@@ -2,7 +2,7 @@
 
 #include <gcs/constraints/all_different.hh>
 #include <gcs/constraints/comparison.hh>
-#include <gcs/constraints/linear_equality.hh>
+#include <gcs/constraints/linear.hh>
 #include <gcs/constraints/not_equals.hh>
 #include <gcs/problem.hh>
 #include <gcs/search_heuristics.hh>
