+++ minicp_benchmarks/tsp/tsp.cc
@@ -1,6 +1,6 @@
 #include <gcs/constraints/circuit.hh>
 #include <gcs/constraints/element.hh>
-#include <gcs/constraints/linear_equality.hh>
+#include <gcs/constraints/linear.hh>
 #include <gcs/problem.hh>
 #include <gcs/search_heuristics.hh>
 #include <gcs/solve.hh>
